+++ src/actions/ActionTypes.tsx
@@ -1,9 +1,11 @@
 // This file serves as a common location for all your action constants
 
-export const CONNECT_SUCCESSFUL = "CONNECT_SUCCESSFUL";
-export const CONNECT_FAILED = "CONNECT_FAILED";
-export const LOGIN_SUCCESSFUL = "LOGIN_SUCCESSFUL";
-export const LOGIN_FAILED = "LOGIN_FAILED";
-export const SOMETHING_HAPPENED_SUCCESSFULLY = "SOMETHING_HAPPENED_SUCCESSFULLY";
-export const SOMETHING_FAILED = "SOMETHING_FAILED";
-export const UPDATE_APICLIENT = "UPDATE_APICLIENT";
+export enum ActionTypes {
+    CONNECT_SUCCESSFUL = "CONNECT_SUCCESSFUL",
+    CONNECT_FAILED = "CONNECT_FAILED",
+    LOGIN_SUCCESSFUL = "LOGIN_SUCCESSFUL",
+    LOGIN_FAILED = "LOGIN_FAILED",
+    SOMETHING_HAPPENED_SUCCESSFULLY = "SOMETHING_HAPPENED_SUCCESSFULLY",
+    SOMETHING_FAILED = "SOMETHING_FAILED",
+    UPDATE_APICLIENT = "UPDATE_APICLIENT",
+}
